Make validate and upload separate actions
Moved some general class to the common directory. Still need to change
the tests to reflect these changes.
